Cleanup changes
* Handling failures in the server.js and exiting in case of failures
* Removing lib.bootstrap call from server.js and also removing it from broker/lib/index.js
* Instead, moving the docker bootstraping call in the start of the docker operator.
* Moving loadCatalogFromAPIServer from broker/lib/index.js to utils
* Removing load of BOSHDirectorClient and Bosh cache in Plan, Changing the get stemcell method
